Red: OrderTests constructor good fail

+++ OrderTracker.Solution/OrderTracker/Models/Order.cs
@@ -6,10 +6,10 @@ namespace OrderTracker.Models
   {
     public string Title {get; set;}
     public string Description { get; set;}
-    public int Price { get; set; }
+    public double Price { get; set; }
     public string Date { get; }
 
-    public Order(string title, string descrip, int price, string date)
+    public Order(string title, string descrip, double price, string date)
     {
       Title = title;
       Description = descrip;
